+++ packages/ui-react/src/components/editor/components/custom-ui/index.tsx
@@ -1,4 +1,5 @@
 import { useEditorAPI } from '@/editor/hooks/useEditorAPI';
+import { LayerDrawer } from '../layer-drawer';
 import { Sidebar } from '../sidebar';
 import { Toolbar } from '../toolbar';
 import { ViewOptions } from '../view-options';
@@ -9,6 +10,7 @@ export const CustomUI = () => {
     <div className="pointer-events-none absolute inset-0 z-[300] flex justify-between [&>*]:pointer-events-auto">
       <Toolbar />
       <ViewOptions />
+      <LayerDrawer />
       <Sidebar />
     </div>
   );