+++ packages/ui-react/src/components/editor/components/sidebar/index.tsx
@@ -1,10 +1,10 @@
 import { useCanvas } from '@/editor/stores/canvas.store';
 import { IconButton } from '@material-tailwind/react';
-import { AnimatePresence, motion } from 'framer-motion';
 import { ArrowRightToLine } from 'lucide-react';
 import { useHotkeys } from 'react-hotkeys-hook';
 import { twMerge } from 'tailwind-merge';
 import { useBoolean } from 'usehooks-ts';
+import { BackDropOverlay } from '../ui/backdrop-overlay';
 import { ReadonlyOverlay } from './readonly-overlay';
 import { SidebarContent } from './sidebar-content';
 
@@ -30,7 +30,7 @@ export const Sidebar = () => {
             'relative h-full w-[calc(70vw)] max-w-xs gap-3 border-l-2 border-gray-900 bg-gray-50 shadow-xl transition-transform duration-300 lg:translate-x-0 lg:duration-0',
             open ? '-translate-x-full' : 'translate-x-0',
           )}>
-          <span className="absolute bottom-10 left-0 -translate-x-full p-6 lg:hidden">
+          <span className="absolute bottom-0 left-0 -translate-x-full p-4 lg:hidden">
             <IconButton onClick={toggleOpen}>
               <ArrowRightToLine
                 className={twMerge(
@@ -46,17 +46,7 @@ export const Sidebar = () => {
         </section>
       </aside>
 
-      <AnimatePresence>
-        {open && (
-          <motion.span
-            onClick={close}
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-            className="absolute left-0 top-0 h-full w-full bg-black/60 backdrop-blur-sm lg:hidden"
-          />
-        )}
-      </AnimatePresence>
+      <BackDropOverlay open={open} onClose={close} />
     </>
   );
 };