+++ packages/ui-react/src/components/editor/components/toolbar/index.tsx
@@ -1,7 +1,7 @@
 import { useEditor } from '@/editor/hooks/useEditor';
 import { ButtonGroup } from '@material-tailwind/react';
 import { useNativeClipboardEvents } from '@tldraw/tldraw';
-import { motion } from 'framer-motion';
+import { AnimatePresence, motion } from 'framer-motion';
 import {
   EraserIcon,
   HighlighterIcon,
@@ -34,94 +34,111 @@ export const Toolbar = () => {
   useNativeClipboardEvents();
 
   return (
-    <motion.div
-      id="editor-toolbar"
-      animate={{ translateY: isReadonly ? '-100%' : 0 }}
-      className="flex h-[min-content] grow flex-col flex-wrap items-center justify-start gap-2 p-4 ease-in-out sm:flex-row sm:items-start sm:justify-center sm:gap-4">
-      <ButtonGroup size="sm" className="order-2">
-        <Action
-          hotkey="alt+s"
-          action={() => setTool('select')}
-          tooltip="Selección">
-          <MousePointer2Icon className="h-4 w-4" />
-        </Action>
-        <Action hotkey="alt+d" action={() => setTool('draw')} tooltip="Dibujar">
-          <PencilIcon className="h-4 w-4" />
-        </Action>
-        <Action
-          hotkey={'alt+l'}
-          variant="text"
-          action={() => setTool('line')}
-          tooltip="Línea">
-          <SlashIcon className="h-4 w-4" />
-        </Action>
-        <Action hotkey="alt+t" action={() => setTool('text')} tooltip="Texto">
-          <TypeIcon className="h-4 w-4" />
-        </Action>
-        <Action
-          hotkey="alt+b"
-          action={() => setTool('eraser')}
-          tooltip="Borrar">
-          <EraserIcon className="h-4 w-4" />
-        </Action>
-      </ButtonGroup>
+    <AnimatePresence initial={false}>
+      {!isReadonly && (
+        <motion.div
+          id="editor-toolbar"
+          initial={{ translateY: '-100%' }}
+          animate={{ translateY: 0 }}
+          exit={{ translateY: '-100%' }}
+          className="flex h-[min-content] grow flex-col flex-wrap items-center justify-start gap-2 p-4 ease-in-out sm:flex-row sm:items-start sm:justify-center sm:gap-4">
+          <ButtonGroup size="sm" className="order-2">
+            <Action
+              hotkey="alt+s"
+              action={() => setTool('select')}
+              tooltip="Selección">
+              <MousePointer2Icon className="h-4 w-4" />
+            </Action>
+            <Action
+              hotkey="alt+d"
+              action={() => setTool('draw')}
+              tooltip="Dibujar">
+              <PencilIcon className="h-4 w-4" />
+            </Action>
+            <Action
+              hotkey={'alt+l'}
+              variant="text"
+              action={() => setTool('line')}
+              tooltip="Línea">
+              <SlashIcon className="h-4 w-4" />
+            </Action>
+            <Action
+              hotkey="alt+t"
+              action={() => setTool('text')}
+              tooltip="Texto">
+              <TypeIcon className="h-4 w-4" />
+            </Action>
+            <Action
+              hotkey="alt+b"
+              action={() => setTool('eraser')}
+              tooltip="Borrar">
+              <EraserIcon className="h-4 w-4" />
+            </Action>
+          </ButtonGroup>
 
-      <ButtonGroup size="sm" className="order-2">
-        <Action
-          hotkey="alt+i"
-          action={() => setTool('asset')}
-          tooltip="Subir archivo">
-          <ImageIcon className="h-4 w-4" />
-        </Action>
-        <Action
-          hotkey="alt+c"
-          action={() => setCustomTool('placeholder-img')}
-          tooltip="Imagen de catálogo">
-          <ReplaceIcon className="h-4 w-4" />
-        </Action>
-        <Action
-          hotkey="alt+s"
-          action={() => setTool('highlight')}
-          tooltip="Subrayador">
-          <HighlighterIcon className="h-4 w-4" />
-        </Action>
-        <Action
-          hotkey="alt+f"
-          action={() => setTool('arrow')}
-          tooltip="Flecha">
-          <LucideArrowUpRight className="h-4 w-4" />
-        </Action>
-        <FreeFormMenu onFormPick={setTool}>
-          <ShapesIcon className="h-4 w-4" />
-        </FreeFormMenu>
-      </ButtonGroup>
+          <ButtonGroup size="sm" className="order-2">
+            <Action
+              hotkey="alt+i"
+              action={() => setTool('asset')}
+              tooltip="Subir archivo">
+              <ImageIcon className="h-4 w-4" />
+            </Action>
+            <Action
+              hotkey="alt+c"
+              action={() => setCustomTool('placeholder-img')}
+              tooltip="Imagen de catálogo">
+              <ReplaceIcon className="h-4 w-4" />
+            </Action>
+            <Action
+              hotkey="alt+s"
+              action={() => setTool('highlight')}
+              tooltip="Subrayador">
+              <HighlighterIcon className="h-4 w-4" />
+            </Action>
+            <Action
+              hotkey="alt+f"
+              action={() => setTool('arrow')}
+              tooltip="Flecha">
+              <LucideArrowUpRight className="h-4 w-4" />
+            </Action>
+            <FreeFormMenu onFormPick={setTool}>
+              <ShapesIcon className="h-4 w-4" />
+            </FreeFormMenu>
+          </ButtonGroup>
 
-      <ButtonGroup variant="outlined" size="sm" className="order-3 sm:order-1">
-        <Action
-          hotkey="ctrl+z"
-          action={() => editor.undo()}
-          disabled={!editor.canUndo || editor.history.numUndos === 1}
-          tooltip="Deshacer">
-          <Undo2Icon className="h-4 w-4" />
-        </Action>
-        <Action
-          hotkey="ctrl+shift+z"
-          action={() => editor.redo()}
-          disabled={!editor.canRedo}
-          tooltip="Rehacer">
-          <Redo2Icon className="h-4 w-4" />
-        </Action>
-        <Action
-          hotkey={['delete', 'backspace']}
-          disabled={editor.selectedShapes.length === 0}
-          action={() => deleteSelected()}
-          tooltip="Borrar">
-          <Trash2Icon className="h-4 w-4" />
-        </Action>
-        <LayerActionsMenu onAction={performAction}>
-          <MoreVerticalIcon className="h-4 w-4" />
-        </LayerActionsMenu>
-      </ButtonGroup>
-    </motion.div>
+          <ButtonGroup
+            variant="outlined"
+            size="sm"
+            className="order-3 sm:order-1">
+            <Action
+              hotkey="ctrl+z"
+              action={() => editor.undo()}
+              disabled={
+                !editor.getCanUndo() || editor.history.getNumUndos() === 1
+              }
+              tooltip="Deshacer">
+              <Undo2Icon className="h-4 w-4" />
+            </Action>
+            <Action
+              hotkey="ctrl+shift+z"
+              action={() => editor.redo()}
+              disabled={!editor.getCanRedo()}
+              tooltip="Rehacer">
+              <Redo2Icon className="h-4 w-4" />
+            </Action>
+            <Action
+              hotkey={['delete', 'backspace']}
+              disabled={editor.getSelectedShapes().length === 0}
+              action={() => deleteSelected()}
+              tooltip="Borrar">
+              <Trash2Icon className="h-4 w-4" />
+            </Action>
+            <LayerActionsMenu onAction={performAction}>
+              <MoreVerticalIcon className="h-4 w-4" />
+            </LayerActionsMenu>
+          </ButtonGroup>
+        </motion.div>
+      )}
+    </AnimatePresence>
   );
 };