+++ packages/ui-react/src/components/editor/components/ui/sortable-context/index.tsx
@@ -1,0 +1,55 @@
+import {
+  DndContext,
+  DragEndEvent,
+  KeyboardSensor,
+  PointerSensor,
+  UniqueIdentifier,
+  useSensor,
+  useSensors,
+} from '@dnd-kit/core';
+import { PropsWithChildren } from 'react';
+
+import { closestCenter } from '@dnd-kit/core';
+import {
+  restrictToFirstScrollableAncestor,
+  restrictToVerticalAxis,
+} from '@dnd-kit/modifiers';
+import {
+  SortableContext,
+  sortableKeyboardCoordinates,
+  verticalListSortingStrategy,
+} from '@dnd-kit/sortable';
+
+type WithId = {
+  id: UniqueIdentifier;
+};
+
+type VerticalSortableContextProps<D extends WithId> = {
+  onDragEnd: (event: DragEndEvent) => void;
+  items: D[];
+};
+
+export const VerticalSortableContext = <D extends WithId>({
+  children,
+  onDragEnd,
+  items,
+}: PropsWithChildren<VerticalSortableContextProps<D>>) => {
+  const sensors = useSensors(
+    useSensor(PointerSensor),
+    useSensor(KeyboardSensor, {
+      coordinateGetter: sortableKeyboardCoordinates,
+    }),
+  );
+
+  return (
+    <DndContext
+      sensors={sensors}
+      collisionDetection={closestCenter}
+      modifiers={[restrictToVerticalAxis, restrictToFirstScrollableAncestor]}
+      onDragEnd={onDragEnd}>
+      <SortableContext items={items} strategy={verticalListSortingStrategy}>
+        {children}
+      </SortableContext>
+    </DndContext>
+  );
+};